+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- Added new command 'Similar' for finding columns that have the same name in other datasets
+- Added the ability to Query Catalogue/DataExport databases directly through RDMP
+
+### Fixed
+
+- AlterColumnType command now shows as IsImpossible when column is part of a view or table valued function
+- Describe command no longer shows relationship properties
+
 ## [6.0.0] - 2021-07-28
 
 ### Changed